refactor(DEV-10629): rename `<I18nLocaleProvider/>` to `<MoniteI18nProvider/>`
